+++ main/rtc_wake_stub_example.h
@@ -22,7 +22,7 @@ typedef struct {
 extern RTC_DATA_ATTR Event event_buffer[];    // Array for storing events
 extern RTC_DATA_ATTR uint32_t s_count;         // Number of events stored
 extern RTC_DATA_ATTR Event lastsent_event;     // Last event sent to MQTT
-extern RTC_DATA_ATTR uint8_t wake_count;       // Number of wakeups (counts up to ten for battery)
+extern RTC_DATA_ATTR uint32_t wake_count;       // Number of wakeups (counts up to ten for battery)
 
 #ifdef __cplusplus
 }
